requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-cors==0.1
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.11.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.0.1
